% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{brnome_freq}
\alias{brnome_freq}
\title{Frenquencia por decada.}
\usage{
brnome_freq(nome, sexo = NULL, localidade_cod = NULL)
}
\arguments{
\item{nome}{Vetor com um ou mais nomes a serem consultados.}

\item{sexo}{Se \code{NULL} (default), a consulta pelo nome eh unissex. Valores podem ser "M", para o sexo masculino, ou "F", para o feminino.}

\item{localidade_cod}{Se \code{NULL} (default), a consulta eh para todo territorio. Para especificar uma localidade, utilize o identificador de uma Unidade da Federação.}
}
\description{
Retorna a frequencia de nascimentos por decada de nascimento para o nome consultado.
}
\examples{
brnome_freq("italo", "m")

}
