% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.R
\name{brnome_rank}
\alias{brnome_rank}
\title{Rankin por sexo, decada ou localidade.}
\usage{
brnome_rank(sexo = NULL, localidade_cod = NULL,
  decada_nascimento = NULL)
}
\arguments{
\item{sexo}{Se \code{NULL} (default), a consulta pelo nome eh unissex. Valores podem ser "M". Para o sexo masculino, ou "F", para o feminino.}

\item{localidade_cod}{Se \code{NULL} (default), a consulta eh para todo territorio. Para especificar uma localidade, utilize o identificador de um município ou de uma Unidade da Federação.}

\item{decada}{Se \code{NULL} (default), a consulta eh para todas as decadas. Para especificar a decada, utilise um numero inteiro.}
}
\description{
Retorna ranking dos nomes segundo a frequencia de nascimentos por sexo, decada ou localidade.
}
\examples{
brnome_rank()

}
