% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localidades.R
\docType{data}
\name{localidades}
\alias{localidades}
\title{Codigo e nome de municipios e estados do brasil}
\format{Tibble}
\source{
\url{https://servicodados.ibge.gov.br/api/docs/localidades?versao=1}
}
\usage{
localidades
}
\description{
Codigo e nome de municipios e estados do brasil
}
\keyword{datasets}
